Refactor: 타입 정의 파일 구조 개선 및 배럴 패턴 적용 - contents 타입 정의를 단일 파일로 통합 - index.ts를 통한 배럴 패턴 도입으로 import 경로 간소화
