+++ src/__mocks__/helpers/fetchers.ts
@@ -1,8 +1,5 @@
 import axios from 'axios';
-import {
-  SectionsResponse,
-  RecommendedResponse,
-} from '@/_types/contents/contents';
+import { SectionsResponse, RecommendedResponse } from '@/_types/contents';
 
 const fetchSections = async (id: number): Promise<SectionsResponse> => {
   const response = await axios.get('/api/sections', {
