+++ src/_types/contents/index.ts
@@ -1,0 +1,1 @@
+export * from './contents.d';
