+++ src/app/contents/main/page.tsx
@@ -2,28 +2,20 @@
 
 import { useEffect, useState } from 'react';
 import Link from 'next/link';
-import {
-  TodayPickContent,
-  BoardSectionSlide,
-} from '@/_types/contents/contents';
+import { TodayPickContent, BoardSectionSlide, Slide } from '@/_types/contents';
 import {
   fetchTodayPick,
   fetchBoardSection,
 } from '@/entities/contents/main/api';
 import { useImageMapping } from '@/entities/contents/hooks';
 
-
 /* Components */
-import {Loading} from '@/features/shared/ui';
-import {StatusBar} from '@/features/shared/ui';
+import { Loading } from '@/features/shared/ui';
+import { StatusBar } from '@/features/shared/ui';
 import { Header } from '@/features/contents/ui/header';
 import { MainSlider } from '@/features/contents/ui/today-pick';
 import { BoardSectionSlider } from '@/features/contents/ui/board-section';
 import { SectionSliderContainer } from '@/features/contents/ui/section-list';
-
-
-/* Types */
-import { Slide } from '@/_types/contents/contents';
 
 /* Styles */
 import styles from './page.module.css';
@@ -114,11 +106,17 @@ const HomePage: React.FC = () => {
       <section>
         <Header headerText={headerText} iconUrl={iconUrl} />
         <MainSlider slides={todayPicks} />
-        <BoardSectionSlider boardSectionSlides={boardSections} getImageSrc={getImageSrc} />
-        <SectionSliderContainer  getImageSrc={getImageSrc}/>
+        <BoardSectionSlider
+          boardSectionSlides={boardSections}
+          getImageSrc={getImageSrc}
+        />
+        <SectionSliderContainer getImageSrc={getImageSrc} />
       </section>
       <div className={styles.recommendContainer}>
-        <img src="/assets/images/dimi-group-text.png" alt="원하는 콘텐츠를 찾지 못하셨나요? Dimi가 직접 추천하는 당신만을 위한 콘텐츠를 확인해보세요." />
+        <img
+          src="/assets/images/dimi-group-text.png"
+          alt="원하는 콘텐츠를 찾지 못하셨나요? Dimi가 직접 추천하는 당신만을 위한 콘텐츠를 확인해보세요."
+        />
         {/* <p>
           <strong>원하는 콘텐츠를 찾지 못하셨나요? </strong> <br />
           <strong>Dimi</strong>가 직접 추천하는 당신만을 위한 콘텐츠를
