+++ src/features/contents/ui/recommended/RecommendedListContainer.tsx
@@ -12,10 +12,7 @@ import { RecommendedSlider } from '@/features/contents/ui/recommended';
 import { fetchRecommended } from '@/entities/contents/recommended';
 
 /* Types */
-import {
-  RecommendedSection,
-  RecommendedResponse,
-} from '@/_types/contents/contents';
+import { RecommendedSection, RecommendedResponse } from '@/_types/contents';
 
 /* Styles */
 import styles from './RecommendedListContainer.module.css';
