+++ src/features/contents/ui/recommended/RecommendedMainSlider.tsx
@@ -4,7 +4,7 @@ import Image from 'next/image';
 
 import Slider from 'react-slick';
 
-import { RecommendedMainSlide } from '@/_types/contents/contents';
+import { RecommendedMainSlide } from '@/_types/contents';
 
 // import 'slick-carousel/slick/slick.css';
 // import 'slick-carousel/slick/slick-theme.css';
