+++ src/features/contents/ui/section-list/SectionSliderContainer.tsx
@@ -13,7 +13,7 @@ import { DropDownOptions } from '@/features/shared/ui/action-bar';
 import { fetchSections, useDropDownStore } from '@/entities/contents/main';
 
 /* Types */
-import { Section, SectionsResponse } from '@/_types/contents/contents';
+import { Section, SectionsResponse } from '@/_types/contents';
 
 /* Styles */
 import styles from './SectionSliderContainer.module.css';
